Тесты, для первого задания черный ящик

+++ lab4/lab4/Program.cs
@@ -5,7 +5,7 @@
 
         static void Main()
         {
-            ShadowLine shadowLine = new ShadowLine(new double[,] { {-2, -3 } });
+            ShadowLine shadowLine = new ShadowLine(new int[,] { {-2, -3 } });
             Console.WriteLine(shadowLine.CalculateSum());
         }
     }
